[refact] add FeedCenter component
