+++ src/app/(MainLayout)/components/Feed/Feed.module.scss
@@ -7,8 +7,3 @@
   border-radius: 15px;
   box-shadow: 3px 3px 10px 0px rgba(0, 0, 0, 0.25);
 }
-
-.feed_center {
-  height: 500px;
-  background-color: rgba(238, 238, 238, 1);
-}
