+++ src/app/(MainLayout)/components/FeedCenter/FeedCenter.module.scss
@@ -1,0 +1,6 @@
+@use '../../../../styles/helpers/index' as *;
+
+.feed_center {
+  height: 500px;
+  background-color: rgba(238, 238, 238, 1);
+}
